修改一些代码

+++ ScrollView_Inner_Slide/src/main/java/com/shenzhen/scrollview_inner_slide/ScrollViewApplication.java
@@ -1,27 +1,24 @@
-package com.shenzhen.recycleview_coupousmoretyple;
+package com.shenzhen.scrollview_inner_slide;
 
 import android.app.Application;
 import android.os.Build;
 import android.os.Environment;
 import android.util.Log;
 
-import org.xutils.x;
-
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.lang.reflect.Field;
 import java.text.DateFormat;
 import java.util.Date;
 
-public class XUtilsApplication extends Application {
+public class ScrollViewApplication extends Application {
 
     @Override
     public void onCreate() {
         super.onCreate();
-        x.Ext.init(this);
-        x.Ext.setDebug(true); //是否输出debug日志，开启debug会影响性能。
 
         Thread.currentThread().setUncaughtExceptionHandler(new MyexceptionHandler());
     }
@@ -31,13 +28,13 @@ public class XUtilsApplication extends Application {
 
         @Override//等发现了为捕获的异常的时候调用的方法；
         public void uncaughtException(Thread thread, Throwable ex) {
-            Log.e("MyexceptionHandler", "......程序发现了异常，被哥们捕获了"+ex);
+            Log.e("MyexceptionHandler", "......程序发现了异常，被哥们捕获了" + ex);
             StringBuffer sb = new StringBuffer();
             Date date = new Date();
             //格式化时间
             String time = DateFormat.getInstance().format(date);
             sb.append("Time:");
-            sb.append(time+"\n");//当前的系统时间；
+            sb.append(time + "\n");//当前的系统时间；
             Field[] fields = Build.class.getDeclaredFields();
            /* for(Field field : fields){
                 try {
@@ -56,15 +53,34 @@ public class XUtilsApplication extends Application {
             sb.append(sw.toString());//追加；
 
             try {  //需要添加读写权限
-                File file = new File(Environment.getExternalStorageDirectory().getPath()+"/Download","Robort_error"+time+".txt");
+                File filedir = new File(Environment.getExternalStorageDirectory().getPath()+"/Robort_error"); //先创建文件夹
+              if (!filedir.exists()) {
+                  filedir.mkdir();
+                }
+                //在创建文件
+                File file = new File(Environment.getExternalStorageDirectory().getPath()+"/Robort_error",System.currentTimeMillis()+".txt");
                 FileOutputStream fos = new FileOutputStream(file);
                 fos.write(sb.toString().getBytes());
                 fos.close();
             } catch (Exception e) {
                 e.printStackTrace();
+                Log.e("MyexceptionHandler", "......程序发现了异常，被哥们捕获了:" + e.getMessage());
             }
+
+//           try {
+//            File mCache = new File(Environment.getExternalStorageDirectory().getPath()+"/", "Robort_error"+System.currentTimeMillis()+ ".txt");
+//            FileOutputStream fos = new FileOutputStream(mCache,false);//创建指定文件; false 代表每次重新写入数据
+//                //创建文件写入流
+//                fos.write(sb.toString().getBytes("gbk"));
+//                fos.flush();
+//                fos.close();
+//            } catch (IOException e) {
+//                e.printStackTrace();
+//                Log.e("MyexceptionHandler", "......程序发现了异常，被哥们捕获了:" + e.getMessage());
+//            }//琵琶
+
        /* android.os.Process.killProcess(android.os.Process.myPid());
-        Log.e("MyexceptionHandler", "......启动自杀方式，再次激活程序");*/
+          Log.e("MyexceptionHandler", "......启动自杀方式，再次激活程序");*/
         }
     }
 }
